+++ client/vite-project/.gitignore
