+++ client/vite-project/lib/main.js
