+++ client/vite-project/package.json
