+++ client/vite-project/src/context/Auth.tsx
