+++ client/vite-project/src/main.tsx
