+++ client/vite-project/src/vite-env.d.ts
