+++ client/vite-project/tsconfig.json
