+++ client/vite-project/vite.config.ts
