+++ docker-compose.yml
@@ -4,39 +4,39 @@ version: "3"
 #   nestjs-node-modules:
 
 services:
-  web:
-    image: vite-image:latest
-    container_name: docker-vite
+  client:
+    image: vite-client-image:latest
+    container_name: client-vite
     working_dir: "/usr/vite-project"
     ports:
       - 5173:5173
     build:
-      context: .
+      context: ./client
       dockerfile: Dockerfile
     # src配下をvolumesとしてmountする
     # https://github.com/docker/compose/issues/8533
     volumes:
       # - src-volume:/usr/vite-project/src
-      - ./vite-project/src/:/usr/vite-project/src/
+      - ./client/vite-project/src/:/usr/vite-project/src/
     restart: always
     tty: true
-    env_file: ./vite-project/.env
-  server:
-    image: vite-server:latest
-    container_name: vite-server
-    working_dir: "/server"
-    ports:
-      - 8080:8080
-    build:
-      context: .
-      dockerfile: Dockerfile_server
-    # src配下をvolumesとしてmountする
-    # https://github.com/docker/compose/issues/8533
-    volumes:
-      # - src-volume:/usr/vite-project/src
-      - ./server/src/server:/server/src/server
-    restart: always
-    tty: true
+    env_file: ./client/vite-project/.env
+  # server:
+  #   image: vite-server:latest
+  #   container_name: vite-server
+  #   working_dir: "/server"
+  #   ports:
+  #     - 8080:8080
+  #   build:
+  #     context: .
+  #     dockerfile: Dockerfile_server
+  #   # src配下をvolumesとしてmountする
+  #   # https://github.com/docker/compose/issues/8533
+  #   volumes:
+  #     # - src-volume:/usr/vite-project/src
+  #     - ./server/src/server:/server/src/server
+  #   restart: always
+  #   tty: true
 
 volumes:
   src-volume:
